requirements.txt
pytesseract
dxcam
python-valve